be calculations format final
