+++ server/src/contracts/contracts.service.ts
@@ -10,7 +10,7 @@ import { createContractErrorHandling } from '../shared/errors/createContractErro
 import * as errorMessages from '../shared/errors/error.messages';
 import { CarsService } from '../cars/cars.service';
 import * as Guard from '../shared/util/Guard';
-import { currentTotalPrice } from '../shared/calculations/priceCalculations';
+import { currentTotalPrice } from '../shared/calculations/prices/currentTotalPrice';
 import { CustomersService } from '../customers/customers.service';
 import { Customer } from '../database/entities/customer.entity';
 import { transformToActiveContractDTO } from './transformers/transformToActiveContractDTO';
