+++ server/src/shared/calculations/days/currentDaysRented.ts
