+++ server/src/shared/calculations/days/estimatedDaysRented.ts
