+++ server/src/shared/calculations/discounts/daysDiscount.spec.ts
