+++ server/src/shared/calculations/prices/estimatedTotalPrice.spec.ts
@@ -1,0 +1,22 @@
+import { getReturnCarData } from '../getReturnedData.spec';
+import { estimatedTotalPrice } from './estimatedTotalPrice';
+
+describe('estimatedTotalPrice', () => {
+    const returnedCarData = getReturnCarData({ phone: 'not applicable' });
+  
+    it('should return the product of estiamtedPricePerDay and estimatedDaysRented', () => {
+      // Arramge
+      const estimatedDaysRentedMock = jest.fn(() => 2);
+      const estiamtedPricePerDayMock = jest.fn(() => 2);
+  
+      // Act
+      const result = estimatedTotalPrice(
+        returnedCarData,
+        estimatedDaysRentedMock,
+        estiamtedPricePerDayMock,
+      );
+  
+      // Assert
+      expect(result).toEqual(4);
+    });
+  });
