+++ server/src/shared/calculations/prices/estimatedTotalPrice.ts
@@ -1,0 +1,13 @@
+import { estimatedPricePerDay } from './estimatedPricePerDay';
+import { estimatedDaysRented } from '../days/estimatedDaysRented';
+
+export const estimatedTotalPrice = (
+    contractData,
+    estimatedPricePerDayFn = estimatedPricePerDay,
+    estimatedDaysRentedFn = estimatedDaysRented,
+  ) => {
+    const estiamtedPricePerDayValue = estimatedPricePerDayFn(contractData);
+    const estimatedDaysRentedNumber = estimatedDaysRentedFn(contractData);
+  
+    return estiamtedPricePerDayValue * estimatedDaysRentedNumber;
+  };
